+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - [FIX] Prevent crash misattribution when an inactive RUM view emits a terminal event after `stopResource()`. See [#2948][]
 - [FEATURE] Instrumented Web Views now have their tracing decision consistent with the native SDK. See [#2859][]
 
+# 3.11.1 / 28-05-2026
+
+- [FIX] Prevent crash in `TracingURLSessionHandler` when wall-clock adjustments produce an inverted start/end timestamp pair during a tracked URL request. See [#2941][]
+- [FIX] Fix `NSInvalidArgumentException` ("unrecognized selector") crash in automatic scroll action tracking when UIKit dispatches a cached delegate method after the original `UIScrollViewDelegate` has been deallocated. See [#2942][].
+- [IMPROVEMENT] Add `trackScrollAndSwipeActions` feature flag in RUM configuration to opt out of automatic scroll and swipe action tracking. See [#2944][].
+
 # 3.11.0 / 12-05-2026
 
 - [IMPROVEMENT] Add support for `maui` source for cross-platform RUM events from .NET MAUI applications. See [#2891][]
@@ -1147,6 +1153,9 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2856]: https://github.com/DataDog/dd-sdk-ios/pull/2856
 [#2866]: https://github.com/DataDog/dd-sdk-ios/pull/2866
 [#2891]: https://github.com/DataDog/dd-sdk-ios/pull/2891
+[#2941]: https://github.com/DataDog/dd-sdk-ios/pull/2941
+[#2942]: https://github.com/DataDog/dd-sdk-ios/pull/2942
+[#2944]: https://github.com/DataDog/dd-sdk-ios/pull/2944
 [#2859]: https://github.com/DataDog/dd-sdk-ios/pull/2859
 [#2948]: https://github.com/DataDog/dd-sdk-ios/pull/2948